backend/requirements.txt
Flask==3.0.2
flask-cors==5.0.0
google-generativeai==0.4.0
python-dotenv==1.0.1
geopy==2.4.1
werkzeug==3.0.1
gunicorn==22.0.0
